Refactor: разделение шаблонов FilmsCard
