+++ src/presenter/films-presenter.js
@@ -2,7 +2,7 @@ import { render } from '../render';
 import FilmsContainer from '../view/films-container-view';
 import FilmsList from '../view/films-list-view';
 import FilmsShowMoreBtn from '../view/films-show-more-btn-view';
-import FilmsCard from '../view/film-card-view';
+import FilmsCard from '../view/film-card';
 import FilmsPopup from '../view/film-popup';
 import { isEsc } from '../util';
