+++ src/view/film-card/index.js
@@ -1,0 +1,27 @@
+import { createElement } from '../../render';
+import createFilmsCardTemplate from './film-card-view';
+
+export default class FilmsCard {
+  #element = null;
+  #film = null;
+
+  constructor(film) {
+    this.#film = film;
+  }
+
+  get template() {
+    return createFilmsCardTemplate(this.#film);
+  }
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+
+    return this.#element;
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
